Convert to single-precision floats

+++ ActivationFunctions/ActivationFunction.h
@@ -3,7 +3,7 @@
 
 class ActivationFunction {
 public:
-    virtual double map(double input) const = 0;
-    virtual double getPreValueDerivative(double input) const = 0;
-    virtual double generateRandomWeight(int32_t connectionsIn, int32_t connectionsOut) const = 0;
+    virtual float map(float input) const = 0;
+    virtual float getPreValueDerivative(float input) const = 0;
+    virtual float generateRandomWeight(int32_t connectionsIn, int32_t connectionsOut) const = 0;
 };
