+++ ActivationFunctions/Relu.h
@@ -3,7 +3,7 @@
 #include "ActivationFunction.h"
 
 class Relu : public ActivationFunction {
-    double map(double input) const;
-    double getPreValueDerivative(double input) const;
-    double generateRandomWeight(int32_t connectionsIn, int32_t connectionsOut) const;
+    float map(float input) const;
+    float getPreValueDerivative(float input) const;
+    float generateRandomWeight(int32_t connectionsIn, int32_t connectionsOut) const;
 };
