+++ CostFunctions/CrossEntropyCost.h
@@ -4,5 +4,5 @@
 class CrossEntropyCost : public CostFunction { // assumes softmax/sigmoid combination
 public:
     using CostFunction::CostFunction;
-    double getCost(Layer& outputLayer, std::vector<int32_t> expected) const;
+    float getCost(Layer& outputLayer, std::vector<int32_t> expected) const;
 };
